python-brain/requirements.txt
Flask
mysql-connector-python
greenswitch
